fix: move BlueprintLookupHint to constraint package to enable serialization
