+++ constraint/blueprint_logderivlookup.go
@@ -1,10 +1,11 @@
-package logderivlookup
+package constraint
 
 import (
 	"fmt"
-
-	"github.com/consensys/gnark/constraint"
 )
+
+// TODO @gbotrel this shouldn't be there, but we need to figure out a clean way to serialize
+// blueprints
 
 // BlueprintLookupHint is a blueprint that facilitates the lookup of values in a table.
 // It is essentially a hint to the solver, but enables storing the table entries only once.
@@ -13,26 +14,11 @@ type BlueprintLookupHint struct {
 }
 
 // ensures BlueprintLookupHint implements the BlueprintSolvable interface
-var _ constraint.BlueprintSolvable = (*BlueprintLookupHint)(nil)
+var _ BlueprintSolvable = (*BlueprintLookupHint)(nil)
 
-// func lookupHint(_ *big.Int, in []*big.Int, out []*big.Int) error {
-// 	nbTable := len(in) - len(out)
-// 	for i := 0; i < len(in)-nbTable; i++ {
-// 		if !in[nbTable+i].IsInt64() {
-// 			return fmt.Errorf("lookup query not integer")
-// 		}
-// 		ptr := int(in[nbTable+i].Int64())
-// 		if ptr >= nbTable {
-// 			return fmt.Errorf("lookup query %d outside table size %d", ptr, nbTable)
-// 		}
-// 		out[i].Set(in[ptr])
-// 	}
-// 	return nil
-// }
-
-func (b *BlueprintLookupHint) Solve(s constraint.Solver, inst constraint.Instruction) error {
+func (b *BlueprintLookupHint) Solve(s Solver, inst Instruction) error {
 	nbEntries := int(inst.Calldata[1])
-	entries := make([]constraint.Element, nbEntries)
+	entries := make([]Element, nbEntries)
 
 	// read the static entries from the blueprint
 	// TODO @gbotrel cache that.
@@ -45,7 +31,7 @@ func (b *BlueprintLookupHint) Solve(s constraint.Solver, inst constraint.Instruc
 	nbInputs := int(inst.Calldata[2])
 
 	// read the inputs from the instruction
-	inputs := make([]constraint.Element, nbInputs)
+	inputs := make([]Element, nbInputs)
 	offset = 3
 	for i := 0; i < nbInputs; i++ {
 		inputs[i], delta = s.Read(inst.Calldata[offset:])
@@ -75,13 +61,13 @@ func (b *BlueprintLookupHint) NbConstraints() int {
 }
 
 // NbOutputs return the number of output wires this blueprint creates.
-func (b *BlueprintLookupHint) NbOutputs(inst constraint.Instruction) int {
+func (b *BlueprintLookupHint) NbOutputs(inst Instruction) int {
 	return int(inst.Calldata[2])
 }
 
 // Wires returns a function that walks the wires appearing in the blueprint.
 // This is used by the level builder to build a dependency graph between instructions.
-func (b *BlueprintLookupHint) WireWalker(inst constraint.Instruction) func(cb func(wire uint32)) {
+func (b *BlueprintLookupHint) WireWalker(inst Instruction) func(cb func(wire uint32)) {
 	return func(cb func(wire uint32)) {
 		// depend on the table UP to the number of entries at time of instruction creation.
 		nbEntries := int(inst.Calldata[1])
@@ -93,7 +79,7 @@ func (b *BlueprintLookupHint) WireWalker(inst constraint.Instruction) func(cb fu
 			n := int(b.EntriesCalldata[j])
 			j++
 			for k := 0; k < n; k++ {
-				t := constraint.Term{CID: b.EntriesCalldata[j], VID: b.EntriesCalldata[j+1]}
+				t := Term{CID: b.EntriesCalldata[j], VID: b.EntriesCalldata[j+1]}
 				if !t.IsConstant() {
 					cb(t.VID)
 				}
@@ -109,7 +95,7 @@ func (b *BlueprintLookupHint) WireWalker(inst constraint.Instruction) func(cb fu
 			n := int(inst.Calldata[j])
 			j++
 			for k := 0; k < n; k++ {
-				t := constraint.Term{CID: inst.Calldata[j], VID: inst.Calldata[j+1]}
+				t := Term{CID: inst.Calldata[j], VID: inst.Calldata[j+1]}
 				if !t.IsConstant() {
 					cb(t.VID)
 				}
